% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/createGrid.R
\name{createGrid}
\alias{createGrid}
\title{Dominated points}
\usage{
createGrid(ep, ngrid)
}
\arguments{
\item{ep,}{ngrid}
}
\value{
hypergrid that contains all intervals
}
\description{
This function is responsible for creating a multidimensional grid
}
\author{
Dennis Assenmacher
}

