% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dominates.R
\name{dominates}
\alias{dominates}
\title{Dominated points}
\usage{
dominates(x, y)
}
\arguments{
\item{x,}{y Vector that holds fitness values.}
}
\value{
TRUE if x dominates y, FALSE if x does not dominate y
}
\description{
Calculates whether a vector x domonates another vector y
}
\examples{
dominates(c(3,2),c(5,6)) == TRUE
}
\author{
Dennis Assenmacher
}

