% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mutate.R
\name{mutate}
\alias{mutate}
\title{PESA-II Mutate}
\usage{
mutate(sol, min, max, sigma = 0.3)
}
\arguments{
\item{min}{Minimum value after mutation}

\item{max}{Maximum value after mutation}

\item{sigma}{Variance}
}
\description{
Gaussian mutation of one individual
}
\author{
Dennis Assenmacher
}

