% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dominates.R
\name{nonDominatedSolutions}
\alias{nonDominatedSolutions}
\title{Non-dominated solutions}
\usage{
nonDominatedSolutions(obj)
}
\arguments{
\item{m}{Matrix or data.frame that contains rowwise entries.}
}
\value{
Matrix which contains all non-dominated points
}
\description{
Calculates all entries which are not dominated by any other entry.
}
\examples{
nonDominatedSolutions(mtcars[,c("mpg","hp")])
}
\author{
Dennis Assenmacher
}

