% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PESAII-package.R, R/PESAII.R
\docType{package}
\name{PESAII}
\alias{PESAII}
\alias{PESAII-package}
\title{PESAII}
\usage{
PESAII(ipMax, epMax, gens, vmin = 0, vmax = 1, nvar = 6, fn, data, nn,
  neighborhoodMutationMatrix, initialPopulation = NULL, crossRate = 0.7,
  sigma = 0.5, nGrid = 20, mock = F, L = 10, printN = NULL)
}
\arguments{
\item{ipMax}{Maximum size of internal solutions used to explore new potential solutions.}

\item{epMax}{Maximum size of external population.}

\item{gens}{Number of generations.}

\item{vmin}{minimum of parameter space variable.}

\item{vmax}{maximum of parameter space variable.}

\item{nvar}{Dimensions of paramaterspace.}

\item{fn}{Multiobjective function that should be optimised.}

\item{data}{Data set as matrix on which to apply \code{PESAII}.}

\item{nn}{Nearest neighbors matrix of \code{data}.}

\item{neighborhoodMutationMatrix}{Neighborhood matrix which mock mutation is based on.}

\item{initialPopulation}{Initial population provided by mock wrapper.}

\item{crossRate}{Probability for parenting.}

\item{sigma}{Standard deviation for mutate function.}

\item{nGrid}{Number of grids.}

\item{L}{Number of nearest neighbors to consider for mutation and mock function.}

\item{printN}{Generation interval in which to print progress of function. Progress bar is displayed, if null.}
}
\description{
FUCK YOU NIGGA

Calculates a list of solutions of a given multiobjective optimisation problem by using the PESA-II algorithm.
}
\author{
Dennis Assenmacher

Dennis Assenmacher
}

