% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/visualizations.R
\name{printParetroFrontPESA}
\alias{printParetroFrontPESA}
\title{PrintCurrentSolutions}
\usage{
printParetroFrontPESA(sol)
}
\arguments{
\item{ep}{External population returned by PESA-II}
}
\description{
Visualizes a 2 dimensional solution of PESA-II
}
\author{
Dennis Assenmacher
}

